Extract shared hook and component from MenuAction and MenuAccordion
Extract useMenuItemRegistration hook for parent context registration
and MenuItemContent component for icon/text rendering.
